chore: rework admin and account related components to use react-query
