+++ src/js/account/__tests__/actions.test.jsx
@@ -1,21 +1,12 @@
-import { GET_ACCOUNT, LOGOUT } from "@app/actionTypes";
+import { GET_ACCOUNT } from "@app/actionTypes";
 import { describe, expect, it } from "vitest";
-import { getAccount, logout } from "../actions";
+import { getAccount } from "../actions";
 
 describe("Account Action Creators:", () => {
     it("getAccount: returns simple action", () => {
         const result = getAccount();
         const expected = {
             type: GET_ACCOUNT.REQUESTED,
-        };
-
-        expect(result).toEqual(expected);
-    });
-
-    it("logout: returns simple action", () => {
-        const result = logout();
-        const expected = {
-            type: LOGOUT.REQUESTED,
         };
 
         expect(result).toEqual(expected);
