--- src/js/administration/mappers.js
@@ -1,3 +1,0 @@
-export const mapSettingsStateToProps = state => ({
-    loading: state.settings.data === null,
-});
