--- src/js/administration/selectors.ts
@@ -1,3 +1,0 @@
-import { get } from "lodash-es";
-
-export const getAccountAdministratorRole = state => get(state, "account.administrator_role", null);
