+++ src/js/base/DropdownMenuLink.tsx
@@ -2,7 +2,15 @@ import { DropdownMenuItem } from "@base/DropdownMenuItem";
 import React from "react";
 import { Link } from "react-router-dom";
 
-export function DropdownMenuLink({ children, className, to, target, rel }) {
+type DropdownMenuLinkProps = {
+    children: React.ReactNode;
+    className?: string;
+    to: string;
+    target?: string;
+    rel?: string;
+};
+
+export function DropdownMenuLink({ children, className, to, target, rel }: DropdownMenuLinkProps) {
     return (
         <DropdownMenuItem className={className} asChild>
             <Link to={to} target={target} rel={rel}>
